Move all major color commands to one script
